requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.0
websockets==12.0
python-dotenv==1.0.1
pyyaml==6.0.2
jinja2==3.1.4
aiosqlite==0.20.0
httpx==0.27.0

# AI
openai==1.50.0

# Voice (Bland.ai)
# Uses httpx for API calls - no SDK needed

# SMS
twilio==9.3.0

# Email
sendgrid==6.11.0

# Scheduling
apscheduler==3.10.4

# Utilities
tenacity==9.0.0
python-json-logger==2.0.7
python-multipart==0.0.9

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
